+++ client/src/pages/RoomsPage.js
@@ -8,13 +8,9 @@ function RoomsPage() {
     const [filteredRooms, setFilteredRooms] = useState([]);
     const [showTable, setShowTable] = useState(false);
     const [showForm, setShowForm] = useState(false);
-    const [patients, setPatients] = useState([]);
-    const [doctors, setDoctors] = useState([]);
 
     useEffect(() => {
         fetchRooms();
-        fetchPatients();
-        fetchDoctors();
     }, []);
 
     const fetchRooms = async () => {
@@ -29,27 +25,6 @@ function RoomsPage() {
         }
     };
 
-    const fetchPatients = async () => {
-        try {
-            const response = await fetch('/patients');
-            if (!response.ok) throw new Error('Network response was not ok');
-            const data = await response.json();
-            setPatients(data);
-        } catch (error) {
-            console.error('Error fetching patient data:', error);
-        }
-    };
-
-    const fetchDoctors = async () => {
-        try {
-            const response = await fetch('/doctors');
-            if (!response.ok) throw new Error('Network response was not ok');
-            const data = await response.json();
-            setDoctors(data);
-        } catch (error) {
-            console.error('Error fetching doctor data:', error);
-        }
-    };
 
 
     
@@ -57,8 +32,6 @@ function RoomsPage() {
     const handleSearch = (searchTerm) => {
         const filtered = rooms.filter(room => {
             // Get the related patient and doctor data
-            const patient = patients.find(p => p.patientID === room.patientID);
-            const doctor = doctors.find(d => d.doctorID === room.doctorID);
             
             // Combine all the fields into a single string to search within
             const combinedData = `
@@ -67,9 +40,6 @@ function RoomsPage() {
                 ${room.number}
                 ${room.occupied}
                 ${room.accommodations}
-                ${room.lengthOfStay}
-                ${patient ? `${patient.firstName} ${patient.lastName} ${patient.patientID}` : ""}
-                ${doctor ? `${doctor.firstName} ${doctor.lastName} ${doctor.doctorID}` : ""}
             `.toLowerCase();
     
             // Check if the search term is present in the combined data
@@ -80,18 +50,15 @@ function RoomsPage() {
     };
 
     
-    
+
     const handleSubmitNewRoom = async (event) => {
         event.preventDefault();
         const form = event.target;
         const newRoom = {
-            patientID: form.patientID.value,
-            doctorID: form.doctorID.value,
             location: form.location.value,
             number: form.number.value,
-            occupied: form.occupied.value === 'true',
+            occupied: form.occupied.value === 'yes',
             accommodations: form.accommodations.value,
-            lengthOfStay: form.lengthOfStay.value,
         };
 
         try {
@@ -120,6 +87,9 @@ function RoomsPage() {
     };
 
     const handleUpdateRoom = async (roomID, updatedRoom) => {
+        const confirmUpdate = window.confirm("Are you sure you want to update this room?");
+        if (!confirmUpdate) return;
+
         try {
             const response = await fetch(`/rooms/${roomID}`, {
                 method: 'PUT',
@@ -143,6 +113,9 @@ function RoomsPage() {
     };
 
     const handleDeleteRoom = async (roomID) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this room?");
+        if (!confirmDelete) return;
+
         try {
             const response = await fetch(`/rooms/${roomID}`, {
                 method: 'DELETE',
@@ -169,8 +142,6 @@ function RoomsPage() {
                     rooms={filteredRooms}
                     onUpdateRoom={handleUpdateRoom}
                     onDeleteRoom={handleDeleteRoom}
-                    patients={patients}
-                    doctors={doctors}
                 />
             </div>
         </>
@@ -179,22 +150,6 @@ function RoomsPage() {
     const renderFormSection = () => (
         <form className="createDataForm" onSubmit={handleSubmitNewRoom}>
             <h3>Add New Room</h3>
-            <select name="patientID">
-                <option value="">Select Patient</option>
-                {patients.map(patient => (
-                    <option key={patient.patientID} value={patient.patientID}>
-                        {patient.patientID} - {patient.firstName} {patient.lastName}
-                    </option>
-                ))}
-            </select>
-            <select name="doctorID">
-                <option value="">Select Doctor</option>
-                {doctors.map(doctor => (
-                    <option key={doctor.doctorID} value={doctor.doctorID}>
-                        {doctor.doctorID} - {doctor.firstName} {doctor.lastName}
-                    </option>
-                ))}
-            </select>
             <select name="location" required>
                 <option value="">Select Location</option>
                 <option value="ICU">ICU</option>
@@ -208,7 +163,6 @@ function RoomsPage() {
                 <option value="No">No</option>
             </select>
             <input type="text" name="accommodations" placeholder="Accommodations" required />
-            <input type="number" name="lengthOfStay" placeholder="Length of Stay (In Days)" />
             <button type="submit" className="btn-action">Submit</button>
         </form>
     );
